feat: Split MainPanel into RoomControlPanel and RoomSearchPanel
- Moved left/right panel UI into separate components
- Ensured consistent global background through BackgroundGradient
- Added navigation from Create button to RoomCreatePanel
